requirements.txt
autopep8
Django
djangorestframework
pandas
psycopg2